rev-expansion-hub/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "rev-expansion-hub"
version = "1.1.2"
description = "High-level asyncio API for REV Robotics Expansion Hubs"
requires-python = ">=3.12"
dependencies = ["rev-core", "rev-rhsplib", "pyserial>=3.5"]

[project.optional-dependencies]
dev = ["mypy"]

[tool.setuptools.packages.find]
where = ["."]
include = ["rev_expansion_hub*"]
